fix(searchForFacetValues): only call client.search (#6465)
This signature is the easiest to detect and makes the required signature of the searchClient simpler (only `search`, `getRecommendations` and `addAlgoliaAgent` are now used).

In this PR I also fix the signature of `search` in `SearchClient` type to make this usage fit reality.

BREAKING CHANGE: in helper.searchForFacetFalues only the method client.search is used, none of the previous fallbacks.