fix(error): directly emit error from events
In `@algolia/events`, errors are properly rethrown if they are thrown with `.emit('error', new Error("msg"))`, not if they are thrown with `.emit('error', { error: new Error('msg') })`. Therefore in InstantSearch and the helper, all errors are now properly thrown directly in the emit.

[FX-3187]

BREAKING CHANGE: if you have any `on('error'` handler that reads `event.error` from the callback, change that to just `error`