feat(routing): change default of cleanUrlOnDispose to false (#6499)
* feat(routing): change default of cleanUrlOnDispose to false

In the vast majority of cases you're either never disposing InstantSearch, or the URL already gets cleared some other way (eg. through SPA navigation). This is why the new default for cleanUrlOnDispose is `false`.

You'd still want to set this option to `true` if you're rendering instantsearch in a modal that should have the url cleared when that modal is closed.

BREAKING CHANGE: default of cleanUrlOnDispose is false

[FX-3213]

* fix tests

* Update packages/instantsearch-core/src/routing/historyRouter.ts

Co-authored-by: Dhaya <154633+dhayab@users.noreply.github.com>

---------

Co-authored-by: Dhaya <154633+dhayab@users.noreply.github.com>