feat(refinementList): add option to disable selecting on searchbox submit (#6649)
* feat(refinementList): add option to disable selecting on searchbox submit

By default when you submit the search box in a refinement list, it toggles the first item. In some cases this is not expected behaviour, so there's now an option `searchableSelectOnSubmit` that defaults to true to allow people to disable this behaviour.

Vue did not have this behaviour, so i didn't implement it in this PR either.

[CR-8634]

* lint