feat(index): introduce "isolated" option
An isolated index does not get requested when changes happen in the parent and also does not inherit its search parameters.

If you want an isolated index to search, you need to cause a search (eg by refining a child widget)

isolated indices:
- do not automatically search when mounted
- do not show up in ui state
- do not interact with  onStateChange
- do not get requested in server-side rendering
- do not cause searches in parent indices

This option is EXPERIMENTAL, and implementation details may change in the future.
Things that could change are:
- which widgets get rendered when a change happens
- whether the index searches automatically
- whether the index is included in the URL / UiState
- whether the index is include in server-side rendering

Usage:

```js
search.addWidgets([
  index({ isolated: true, indexName: 'alone' })
]);

search.addWidgets([
  index({ isolated: false }).addWidgets([
    index({ indexName: 'isolated' })
  ]),
]);
```

[FX-3448]

supersedes the following POCs:
- closes #6880
- closes #5939