fix(perf): don't render loading icon if not visible
In Safari, an animation with animateTransform being in the dom (even if hidden) will cause a style recalculation on every frame. This also causes needless processor usage, and slowdown on old devices.

This PR solves this issue in instantsearch by preventing rendering of the loading icon if it shouldn't be displayed.

- react: remove rendering
- js: remove rendering
- vue: already not rendering

ref: https://bugs.webkit.org/show_bug.cgi?id=298217
ref: https://github.com/algolia/autocomplete/issues/1322
ref: https://github.com/algolia/autocomplete/pull/1323