fix(deps): depend on react (#6753)
* fix(deps): depend on react

This is a temporary fix for https://github.com/quantizor/markdown-to-jsx/issues/716, as there's an import for react in markdown-to-jsx that causes parcel to try to evaluate react. If there's no react dependency it throws.

We add a unused dependency on react, until either markdown-to-jsx has an entry point that does not depend on react existing (it's not actually used) or we fork it.

* fix